requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.3.0
    # via anyio
fastapi[all]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.3
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   openai
huggingface-hub==0.32.4
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
jiter==0.10.0
    # via openai
llvmlite==0.44.0
    # via numba
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numba==0.61.2
    # via openai-whisper
numpy==2.2.6
    # via
    #   -r requirements.in
    #   numba
    #   openai-whisper
    #   soundfile
    #   transformers
openai==1.84.0
    # via -r requirements.in
openai-whisper==20240930
    # via -r requirements.in
orjson==3.10.18
    # via fastapi
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pycountry==24.6.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   fastapi
    #   openai
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.5
    # via fastapi
pydantic-settings==2.9.1
    # via fastapi
pydub==0.25.1
    # via -r requirements.in
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via
    #   fastapi
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   tiktoken
    #   transformers
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
safetensors==0.5.3
    # via transformers
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soundfile==0.13.1
    # via -r requirements.in
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tiktoken==0.9.0
    # via openai-whisper
tokenizers==0.21.1
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   openai-whisper
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   openai-whisper
    #   transformers
transformers==4.52.4
    # via -r requirements.in
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.14.0
    # via
    #   anyio
    #   exceptiongroup
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich
    #   rich-toolkit
    #   torch
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.10.0
    # via fastapi
urllib3==2.4.0
    # via requests
uvicorn[standard]==0.34.3
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
